refactor: adapter 패키지 세부화
