+++ src/test/java/yapp/buddycon/app/auth/JwtTokenProviderTest.java
@@ -5,9 +5,9 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import yapp.buddycon.app.auth.adapter.JwtTokenCreator;
-import yapp.buddycon.app.auth.adapter.infra.RedisRefreshTokenStorage;
-import yapp.buddycon.app.auth.adapter.JwtTokenProvider;
+import yapp.buddycon.app.auth.adapter.jwt.JwtTokenCreator;
+import yapp.buddycon.app.auth.adapter.redis.RedisRefreshTokenStorage;
+import yapp.buddycon.app.auth.adapter.jwt.JwtTokenProvider;
 import yapp.buddycon.app.auth.application.service.LocalTime;
 import yapp.buddycon.app.auth.application.service.Token;
 import yapp.buddycon.app.user.domain.User;
